backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
langgraph
langchain
langchain-community
langchain-core
pydantic>=2.0
qdrant-client
open_clip_torch
torch
torchvision
sqlalchemy
pillow
pydantic-settings
pymupdf
python-multipart
python-dotenv
httpx
langchain-google-genai>=1.0.3
langchain-groq
google-generativeai>=0.5.2
langchain-text-splitters
tenacity
cachetools
passlib[bcrypt]
python-jose[cryptography]
cryptography
bcrypt==4.0.1
email-validator>=2.0.0
